Move input.hpp into game_logic_common
