+++ src/frontend/input_handler.hpp
@@ -18,7 +18,7 @@
 
 #include "base/spatial_types.hpp"
 #include "base/warnings.hpp"
-#include "game_logic/input.hpp"
+#include "game_logic_common/input.hpp"
 
 RIGEL_DISABLE_WARNINGS
 #include <SDL.h>
