+++ src/game_logic/camera.hpp
@@ -19,8 +19,8 @@
 #include "base/spatial_types.hpp"
 #include "base/warnings.hpp"
 #include "engine/base_components.hpp"
-#include "game_logic/input.hpp"
 #include "game_logic/player/components.hpp"
+#include "game_logic_common/input.hpp"
 
 RIGEL_DISABLE_WARNINGS
 #include <entityx/entityx.h>
