+++ src/game_logic/game_world.hpp
@@ -28,8 +28,8 @@
 #include "frontend/game_mode.hpp"
 #include "game_logic/damage_components.hpp"
 #include "game_logic/global_dependencies.hpp"
-#include "game_logic/input.hpp"
 #include "game_logic_common/igame_world.hpp"
+#include "game_logic_common/input.hpp"
 #include "ui/hud_renderer.hpp"
 #include "ui/ingame_message_display.hpp"
 #include "ui/menu_element_renderer.hpp"
