+++ src/game_logic/global_dependencies.hpp
@@ -21,7 +21,7 @@
 #include "base/warnings.hpp"
 #include "data/tutorial_messages.hpp"
 #include "engine/base_components.hpp"
-#include "game_logic/input.hpp"
+#include "game_logic_common/input.hpp"
 
 RIGEL_DISABLE_WARNINGS
 #include <entityx/entityx.h>
