+++ src/game_logic/player/interaction_system.hpp
@@ -22,8 +22,8 @@
 #include "data/player_model.hpp"
 #include "data/tutorial_messages.hpp"
 #include "engine/base_components.hpp"
-#include "game_logic/input.hpp"
 #include "game_logic/player/components.hpp"
+#include "game_logic_common/input.hpp"
 
 RIGEL_DISABLE_WARNINGS
 #include <entityx/entityx.h>
